feat: update the "command palette" extension point to use the hook
