+++ public/app/features/commandPalette/actions/useActions.ts
@@ -6,18 +6,20 @@ import { CommandPaletteAction } from '../types';
 
 import { getRecentDashboardActions } from './dashboardActions';
 import getStaticActions from './staticActions';
+import useExtensionActions from './useExtensionActions';
 
 export default function useActions(searchQuery: string) {
   const [navTreeActions, setNavTreeActions] = useState<CommandPaletteAction[]>([]);
   const [recentDashboardActions, setRecentDashboardActions] = useState<CommandPaletteAction[]>([]);
+  const extensionActions = useExtensionActions();
 
   const navBarTree = useSelector((state) => state.navBarTree);
 
   // Load standard static actions
   useEffect(() => {
-    const staticActionsResp = getStaticActions(navBarTree);
+    const staticActionsResp = getStaticActions(navBarTree, extensionActions);
     setNavTreeActions(staticActionsResp);
-  }, [navBarTree]);
+  }, [navBarTree, extensionActions]);
 
   // Load recent dashboards - we don't want them to reload when the nav tree changes
   useEffect(() => {
